compiled jar file, and final fixes
added the compiled versiion to the source controll.
there are some problems example the indexing of units is controlled in the client class, witch has to be fixed. also the client class aint a client, do to time constrains.

the core functionality is tested in unit tests, but in later versions i need to test more indeapth.